changed all references using millis to unsigned long to fit the output of the function refactored fan and other minor changes.

+++ firmware/lib/core/include/logger.h
@@ -30,7 +30,7 @@ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
 namespace CogCore {
 
 struct LogRecord {
-    int timestamp;
+    unsigned long timestamp;
     LogLevel level;
     LogMessage message;
 };
